fpsensor: Fix g++ compiler error
When compiling the host tests, g++ warns about the following:

common/fpsensor/fpsensor.cc:491:25: error: comparison of integer
expressions of different signedness: 'uint32_t' {aka 'unsigned int'} and
'int8_t' {aka 'signed char'} [-Werror=sign-compare]
  491 |                 if (fgr == template_newly_enrolled) {
      |                     ~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~

BRANCH=none
BUG=b:234181908, b:244781166
TEST=make host-fpsensor_crypto
TEST=./util/compare_build.sh -b fp -j 120
     => MATCH

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3689786

+++ common/fpsensor/fpsensor.c
@@ -481,7 +481,11 @@ static enum ec_status fp_command_frame(struct host_cmd_handler_args *args)
 				FP_CONTEXT_ENCRYPTION_SALT_BYTES);
 		trng_exit();
 
-		if (fgr == template_newly_enrolled) {
+		/*
+		 * TODO(http://b/244781166): Use consistent types so cast is
+		 * not needed.
+		 */
+		if (fgr == (uint32_t)template_newly_enrolled) {
 			/*
 			 * Newly enrolled templates need new positive match
 			 * salt, new positive match secret and new validation
